+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - widgets
   - frames: allow loading local html files without requiring a web server to serve them
+  - text: escape html tags characters (`<` & `>`)
 
 ## 0.47.4