+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - bug fixes
   - package: broken build from sources on Windows
   - editor: incorrect position when moving a widget from an edge with the keyboard
+  - modal: prevent touchend event from reaching the inner widgets when opening the container
 
 - widgets
   - frames: allow loading local html files without requiring a web server to serve them