+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - matrix: clone matrices not properly storing their children's values
 
 - widgets
+  - **new** html widget
   - frames: allow loading local html files without requiring a web server to serve them
   - text: escape html tags characters (`<` & `>`)
   - input: added `asYouType` property