+++ CHANGELOG.md
@@ -1,14 +1,12 @@
 # Changelog
 
-## 0.47.5
-
+## 0.48.0
 
 - bug fixes
   - package: broken build from sources on Windows
   - editor: incorrect position when moving a widget from an edge with the keyboard
   - modal: prevent touchend event from reaching the inner widgets when opening the container
   - matrix: clone matrices not properly storing their children's values
-
 - widgets
   - **new** html widget
   - frames: allow loading local html files without requiring a web server to serve them